+++ src/endpoints/quotes/quotes-service.js
@@ -11,20 +11,22 @@ const quotesService = {
       .then(rows => rows[0]);
   },
 
-  updateQuote(db, updatedQuote) {
+  updateQuote(db, updatedQuote, accountId) {
     return db
       .update(updatedQuote)
       .into('customer_quotes')
       .where('customer_quote_id', '=', updatedQuote.customer_quote_id)
+      .andWhere('account_id', accountId)
       .returning('*')
       .then(rows => rows[0]);
   },
 
-  deleteQuote(db, quoteID) {
+  deleteQuote(db, quoteID, accountId) {
     return db
       .delete()
       .from('customer_quotes')
       .where('customer_quote_id', '=', quoteID)
+      .andWhere('account_id', accountId)
       .returning('*')
       .then(rows => rows[0]);
   }
